salary details

+++ HRMS_Core/Salary/SalaryDetailViewModel.cs
@@ -41,7 +41,7 @@ namespace HRMS_Core.Salary
         public decimal? PresentDays { get; set; }
         public int? WeekOff { get; set; }
         public decimal? Leave { get; set; }
-        public int? SalaryDays { get; set; }
+        public decimal? SalaryDays { get; set; }
         public decimal? Arrears { get; set; }
         public decimal? OtherDeduction { get; set; }
         public string? BranchName { get; set; }
